fix(tsconfig): inherit base module settings
