+++ packages/accounts/tsconfig.json
@@ -1,17 +1,13 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "module": "NodeNext",
     "outDir": "dist",
     "rootDir": "src",
     "declaration": true,
     "declarationMap": true,
     "allowImportingTsExtensions": false,
     "emitDeclarationOnly": false,
-    "types": ["node", "vitest/globals"],
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "moduleResolution": "node"
+    "types": ["node", "vitest/globals"]
   },
   "include": ["src", "types/**/*.d.ts"]
 }
