+++ packages/reporting/tsconfig.json
@@ -5,10 +5,7 @@
     "rootDir": ".",
     "declaration": true,
     "declarationMap": true,
-    "types": ["vitest/globals"],
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "moduleResolution": "nodenext"
+    "types": ["vitest/globals"]
   },
   "include": ["src", "types/**/*.d.ts", "vitest.config.ts"]
 }
