+++ packages/rules/tsconfig.json
@@ -4,9 +4,7 @@
     "outDir": "dist",
     "module": "ESNext",
     "moduleResolution": "Bundler",
-    "types": ["vitest/globals"],
-    "resolveJsonModule": true,
-    "skipLibCheck": true
+    "types": ["vitest/globals"]
   },
   "include": ["__tests__", "src", "types/**/*.d.ts"]
 }
