+++ packages/strategy-apx-ddb01/tsconfig.json
@@ -1,14 +1,12 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "target": "ES2022",
     "module": "ES2022",
     "moduleResolution": "Bundler",
     "outDir": "dist",
     "declaration": true,
     "declarationMap": false,
     "sourceMap": false,
-    "strict": true,
     "rootDir": "."
   },
   "include": ["src/**/*", "test/**/*"]
